fix test_samplers

+++ tests/dataloaders/test_dataloaders.py
@@ -131,31 +131,31 @@ def test_anndataloader_distributed_sampler(num_processes: int, save_path: str):
 # @pytest.mark.optional
 @pytest.mark.parametrize("num_processes", [1, 2])
 def test_scanvi_with_distributed_sampler(num_processes: int, save_path: str):
-    #if torch.cuda.is_available():
-    adata = scvi.data.synthetic_iid()
-    manager = generic_setup_adata_manager(adata)
-    SCANVI.setup_anndata(
-        adata,
-        "labels",
-        "label_0",
-        batch_key="batch",
-    )
-    file_path = save_path + "/dist_file"
-    if os.path.exists(file_path):  # Check if the file exists
-        os.remove(file_path)
-    datasplitter_kwargs = {}
-    datasplitter_kwargs["distributed_sampler"] = True
-    if num_processes == 1:
-        datasplitter_kwargs["distributed_sampler"] = False
-    datasplitter_kwargs["save_path"] = save_path
-    datasplitter_kwargs["num_processes"] = num_processes
-    model = SCANVI(adata, n_latent=10)
+    if torch.cuda.is_available():
+        adata = scvi.data.synthetic_iid()
+        manager = generic_setup_adata_manager(adata)
+        SCANVI.setup_anndata(
+            adata,
+            "labels",
+            "label_0",
+            batch_key="batch",
+        )
+        file_path = save_path + "/dist_file"
+        if os.path.exists(file_path):  # Check if the file exists
+            os.remove(file_path)
+        datasplitter_kwargs = {}
+        datasplitter_kwargs["distributed_sampler"] = True
+        if num_processes == 1:
+            datasplitter_kwargs["distributed_sampler"] = False
+        datasplitter_kwargs["save_path"] = save_path
+        datasplitter_kwargs["num_processes"] = num_processes
+        model = SCANVI(adata, n_latent=10)
 
-    torch.multiprocessing.spawn(
-        multiprocessing_worker,
-        args=(num_processes, manager, save_path),
-        nprocs=num_processes,
-        join=True,
-    )
+        torch.multiprocessing.spawn(
+            multiprocessing_worker,
+            args=(num_processes, manager, save_path),
+            nprocs=num_processes,
+            join=True,
+        )
 
-    model.train(1, datasplitter_kwargs=datasplitter_kwargs)
+        model.train(1, datasplitter_kwargs=datasplitter_kwargs)
